% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{infuse}
\alias{infuse}
\title{Infuse a template with values.}
\usage{
infuse(file_or_string, ..., variable_identifier = c("{{", "}}"),
  default_char = "|", transform_function = function(value) return(value),
  verbose = FALSE)
}
\arguments{
\item{file_or_string}{the template file or a string containing the template}

\item{...}{different keys with related values, used to fill in the template}

\item{variable_identifier}{the opening and closing character that denounce a variable in the template}

\item{default_char}{the character use to specify a default after}

\item{transform_function}{a function through which all specified values are passed, can be used to make inputs safe(r).  dplyr::build_sql is a good default for SQL templating.}

\item{verbose}{verbosity level}
}
\description{
To help prevent \href{https://xkcd.com/327/}{SQL injection attacks} (or other injection attacks), use a transformation function to escape special characters. \code{\link[dplyr]{build_sql}} is a great default escaping function for SQL templating.  For templating in other languages you will need to build your own escaping function.
}

