% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.infuse}
\alias{print.infuse}
\title{prints/shows the result of the \code{infuse} function using the \code{cat} function}
\usage{
\method{print}{infuse}(x, ...)
}
\arguments{
\item{x}{output of the \code{infuse} function}
}
\description{
prints/shows the result of the \code{infuse} function using the \code{cat} function
}

