% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_template}
\alias{read_template}
\title{Reads the template from a file or string}
\usage{
read_template(file_or_string)
}
\arguments{
\item{file_or_string}{the file or character string to read}
}
\description{
Reads the template from a file or string
}

