% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim}
\alias{trim}
\title{Returns string w/o leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{to string to strip}
}
\description{
Returns string w/o leading or trailing whitespace
}

