% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/core.R
\name{variables_requested}
\alias{variables_requested}
\title{Shows which variables are requested by the template}
\usage{
variables_requested(file_or_string, variable_identifier = c("{{", "}}"),
  default_char = "|", verbose = FALSE)
}
\arguments{
\item{file_or_string}{the template file or a string containing the template}

\item{variable_identifier}{the opening and closing character that denounce a variable in the template}

\item{default_char}{the character use to specify a default after}

\item{verbose}{verbosity level}
}
\description{
Shows which variables are requested by the template
}

