% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core.R
\name{infuse}
\alias{infuse}
\title{Infuse a template with values}
\usage{
infuse(file_or_string, ..., variable_identifier = c("{{", "}}"),
  default_char = "|", verbose = FALSE)
}
\arguments{
\item{file_or_string}{the template file or a string containing the template}

\item{...}{different keys with related values, used to fill in the template}

\item{variable_identifier}{the opening and closing character that denounce a variable in the template}

\item{default_char}{the character use to specify a default after}

\item{verbose}{verbosity level}
}
\description{
Infuse a template with values
}

