% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.infuse}
\alias{print.infuse}
\title{prints/shows the result of the \code{infuse} function using the \code{cat} function}
\usage{
\method{print}{infuse}(x, ...)
}
\arguments{
\item{x}{output of the \code{infuse} function}

\item{...}{further arguments passed to or from other methods.}
}
\description{
prints/shows the result of the \code{infuse} function using the \code{cat} function
}

